Allow plot selection
Add a Plot Type dropdown box to the UI which allows the user to select
the plot to use to display the data. Change the date processing in
`server.R` to expect datetimes instead of dates.

This allows more flexibility for the user and opens the way to more
kinds of plots.