Create a reference plot to compare time selections
Create a third graph that plots data in the time period selected in the
two primary plots (currently only deviation cost is supported, and so
the bar chart graph is not used). The individual plots are colour coded.
When the two time period selections have different durations, the data
is not scaled, just plotted as-is, so one line may be much shorter than
the other.

This allows a primitive comparison of user-selected time periods.