Stop filtering by measure in data tidying
The `convert()` functions in `read.R` will no longer only operate on
columns in the original data with the mass flow cost measure, but
instead expand every single column. The resulting data file is quite
large but contains information about every measure. The purpose of this
is to eventually allow the user to filter by measure in the UI.