Allow filtered selection of deviated variable
Add dropdown menus to the UI that allow the deviated variable to be
selected from a reduced list after choosing the unit and the type
(currently blr, turb, efficiency).

The measure can also be selected but is not functional.

Requires running `regenerate()` in `read.R` in order to create
`columns.yaml`.