Allow selection of measure for deviated variable
Selecting a measure in the dropdown box for a deviated variable will now
update the plots and statistics accordingly.

`u$extract()` has been changed to take both a variable name and a
measure name.

Currently there is no restriction on comparing two deviated variables
with different measures selected. This may allow for meaningless
combinations to be chosen.