Create comparative plot for timeline selections
Create a second plot on the single-component "Across Time" tab which
graphs both of the user's selections on the timeline plot together. This
is useful for comparing two graph features in the lifetime of a
particular component.

Note that only the left component selectors are currently used for this
tab, and the right ones are ignored.