Add toggle checkbox and aggregated cost stat
Allow the right component to be toggled on/off via a checkbox. The UI
will show either the "two component graphs with reference graph" or the
"one timeline graph and one comparative graph" layout accordingly, with
no need to switch tabs.

Add an aggregated cost statistic to both the reference and timeline
comparative plots as a legend. Code that generates legends for different
statistics is located in `stats.R`.

This commit breaks the "Bar Chart" comparative plots. Only the "Line
Chart" plots implement the new functionality.